refactor(bridge): extract SettingsMessageHandler
